+++ src/features/items/components/ItemCard.tsx
@@ -89,7 +89,6 @@ export function ItemCard({ data: item, onMoreInfoClick }: Props) {
           {(MutatorItem.isMutatorItem(item) || TraitItem.isTraitItem(item)) && (
             <div className="mt-3 whitespace-pre-line text-left text-xs text-gray-200">
               <strong>At Max Level: </strong>
-              {item.maxLevelBonus || 'No max level bonus found.'}
               <DescriptionWithTags
                 description={item.maxLevelBonus || 'No max level bonus found.'}
               />