+++ src/features/items/data/remnantItems.ts
@@ -4715,7 +4715,7 @@ export const remnantItems: Item[] = [
     name: 'Corrupted Nebula',
     imagePath: '/offhand/corrupted_nebula.png',
     id: '8u8j4G',
-    description: `Semi-automatic laser weapon. Weakspot hits and kills create a Nano-bot for 8s that seeks enemies within 20m and repeatedly attacks dealing 8.4 ACID damage per hit.Max 5 Nano-bots.`,
+    description: `Semi-automatic laser weapon. Weakspot hits and kills create a Nano-bot for 8s that seeks enemies within 20m and repeatedly attacks dealing 6-12 ACID damage per hit. Max 5 Nano-bots.`,
     wikiLinks: ['https://remnant.wiki/Corrupted_Nebula'],
     linkedItems: {
       mod: {
@@ -5776,9 +5776,7 @@ export const remnantItems: Item[] = [
     name: 'Nano Phase',
     imagePath: '/mod/nano_phase.png',
     id: 'WARGv3',
-    description: `Detonate active Nano-bots, each dealing 100 ACID damage [E] plus an additional 25 ACID damage per Nano-bot detonated to enemies within 2m and applies CORRODED, dealing 50 damage [E] over 10s per Nano-bot detonated.
-    
-    Mod Power Requirement: 1,000`,
+    description: `Detonate active Nano-bots, each dealing 100 - 200 ACID damage [E] plus an additional 25 - 50 ACID damage per Nano-bot detonated to enemies within 2m and applies CORRODED, dealing 50 - 100 damage [E] over 10s per Nano-bot detonated.`,
     linkedItems: {
       weapon: {
         name: 'Corrupted Nebula',